Restructure animated map components
